refactor: GNBItem 컴포넌트 분리
