+++ src/components/Header/index.tsx
@@ -1,34 +1,10 @@
 import { useState } from 'react';
 
 import * as S from './style';
-import { IGNBItem } from './types';
 
 import Icon from '@/components/common/Icons';
+import GNBItem from '@/components/Header/GNBItem';
 import { GNB } from '@/constants/gnb';
-
-const GNBItem = ({ name, src, icon }: IGNBItem) => {
-  return (
-    <S.Item
-      fontWeight='medium'
-      fontSize='md'
-      lineHeight='tall'
-      letterSpacing='wide'
-      $tColor='blue 600'
-      $bColor='gray 400'>
-      <S.NLink
-        $tColor='blue 600'
-        $bColor='gray 400'
-        to={src}
-        flex='flex'
-        $justifyContent='center'
-        $alignItems='center'
-        className={({ isActive }) => (isActive ? 'active' : undefined)}>
-        {icon}
-        {name}
-      </S.NLink>
-    </S.Item>
-  );
-};
 
 const Header = () => {
   const [isShowing, setIsShowing] = useState(false);
